Update .bashrc file to improve the terminal prompt via PS1 var!
Now, instead of just getting this, for instance:
```
$SHLVL:1 master
gs-hp-spectre ~/dev/eRCaGuy_dotfiles $
```
We get this! Notice the addition of the username, rather than just
having the hostname. Also, it shows it all *in color* now, which is
the default for a standard Ubuntu 18 installation, so that behavior
is no longer lost like it was before!
```
$SHLVL:1 master
gabriel@gs-hp-spectre:~/dev/eRCaGuy_dotfiles$
```

See this image for more a more detailed and full-color look:
"bashrc_sample_terminal_prompt.png"