add apt-cacher_server_proxy module to check status & toggle ON/OFF
This module allows one to check to see if an apt-cacher server is in
use, and to toggle its usage ON or OFF if desired.

This is useful, for example, if installing via `sudo apt-get install`
or `sudo apt install` while on a certain network, versus not. On a
given private network you may have an apt-cacher proxy server, for
example, which you want to use, but when off this network you just want
to use the public apt server instead of your private apt-cacher server.
This script allows you to easily toggle the private server usage on or
off.

- See instructions in each file for "installation" & usage.

Files added:

useful_scripts/apt-cacher-server_proxy.sh
useful_scripts/apt-cacher-server_proxy_status.sh
useful_scripts/apt-cacher-server_proxy_toggle.sh