git-diffn.sh: fix various bugs; this is the FIRST WORKING VERSION!
The most difficult bug to find was adding the `\033` in the regexp
list of chars to exclude, near the far right, here!

    match(raw_line, /^(\033\[[0-9]{1,2}m)?(---|\+\+\+|[^-+ \033])/) {

The next most-important and difficult-to-find bug fix was escaping the
+ char near the far right here:

    if (raw_line ~ /^(\033\[[0-9]{1,2}m)?\+/) {

And lastly, was fixing a few bad `printf` calls with swapped usages of
`left_num` and `right_num`.