useful_scripts/sync_git_repo_from_pc1_to_pc2.sh: print SUMMARY at end...
...which shows commit hash synced, From PC, and To PC.

This makes it MUCH easier to verify the receiving PC actually has the
correct  commit in its log, by running

    git log -1

to view the latest log on it! Now you can easily compare its commit
has with the one from the SUMMARY, which was just synced, to make sure
they match.